requirements.txt
flask
dotenv
gunicorn
apscheduler==3.11.0
gitpython
